Add -p arg to preserve commit hash references
This option delegates the creation of the final repo to
git-filter-repo instead of fast-import. git-filter-repo has
fascilities to update references to commit hashes in
commit messages to the hash of the new commit, so all
cherry-pick references remain meaningful. To do so it requires
that the exports have the 'original-oid' in the exported file.

Changes here add support for the 'original-oid' in the fast-export
data and delegating the import to git-filter-repo. Note that
using this option requires the user have git-filter-repo.