Removes remapping of tags to new marks
Tags are a special case and should not have their parent mark
remapped. With the current arguments to fast-export (which don't
include --mark-tags for example) tags are formatted like this:

tag 1.5.0
from :3493
original-oid eb12ca36ae2a859ba1ae262e1c9153b062940140

The previous implementation called
`commands.append(remapmark(src_commands[k], mark_map))` for every
line of the fast-export. To avoid remapping the tag "from" marks, a
state variable is introduced to track if a "tag from mark" is expected,
in which case the command is appended without remapping.

Also of note, the tags are all placed at the end of the fast-export
stream, so this special case handling is done after the logs are
merged, when the tail end of the commands are handled.